# specs/002-todo-app-web/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo App Phase II API
  description: REST API for the Todo App Phase II - a full-stack web application with authentication
  version: 1.0.0
  contact:
    name: Todo App Support
    email: support@todoapp.com

servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.todoapp.com
    description: Production server

paths:
  /api/{user_id}/tasks:
    get:
      summary: Get all tasks for a user
      description: Retrieve all tasks associated with the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user whose tasks to retrieve
          schema:
            type: string
        - name: completed
          in: query
          required: false
          description: Filter tasks by completion status
          schema:
            type: boolean
      responses:
        '200':
          description: List of tasks for the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []

    post:
      summary: Create a new task
      description: Create a new task for the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user creating the task
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - bearerAuth: []

  /api/{user_id}/tasks/{id}:
    get:
      summary: Get a specific task
      description: Retrieve a specific task by ID for the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user who owns the task
          schema:
            type: string
        - name: id
          in: path
          required: true
          description: The ID of the task to retrieve
          schema:
            type: integer
      responses:
        '200':
          description: The requested task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - bearerAuth: []

    put:
      summary: Update a task
      description: Update an existing task for the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user who owns the task
          schema:
            type: string
        - name: id
          in: path
          required: true
          description: The ID of the task to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - bearerAuth: []

    delete:
      summary: Delete a task
      description: Delete a specific task for the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user who owns the task
          schema:
            type: string
        - name: id
          in: path
          required: true
          description: The ID of the task to delete
          schema:
            type: integer
      responses:
        '204':
          description: Task deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - bearerAuth: []

  /api/{user_id}/tasks/{id}/complete:
    patch:
      summary: Toggle task completion status
      description: Update the completion status of a specific task
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user who owns the task
          schema:
            type: string
        - name: id
          in: path
          required: true
          description: The ID of the task to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCompleteRequest'
      responses:
        '200':
          description: Task completion status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Task:
      type: object
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Unique identifier for the task
          example: 1
        user_id:
          type: string
          description: ID of the user who owns the task
          example: "user_abc123"
        title:
          type: string
          description: Title of the task
          minLength: 1
          maxLength: 255
          example: "Buy groceries"
        description:
          type: string
          description: Optional description of the task
          maxLength: 1000
          example: "Milk, eggs, bread"
        completed:
          type: boolean
          description: Whether the task is completed
          example: false
        created_at:
          type: string
          format: date-time
          description: Timestamp when the task was created
          example: "2026-01-23T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the task was last updated
          example: "2026-01-23T10:00:00Z"

    TaskCreateRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: Title of the task
          minLength: 1
          maxLength: 255
          example: "Buy groceries"
        description:
          type: string
          description: Optional description of the task
          maxLength: 1000
          example: "Milk, eggs, bread"
        completed:
          type: boolean
          description: Whether the task is completed (defaults to false)
          example: false

    TaskUpdateRequest:
      type: object
      required:
        - title
        - description
        - completed
      properties:
        title:
          type: string
          description: Title of the task
          minLength: 1
          maxLength: 255
          example: "Buy groceries"
        description:
          type: string
          description: Optional description of the task
          maxLength: 1000
          example: "Milk, eggs, bread"
        completed:
          type: boolean
          description: Whether the task is completed
          example: false

    TaskCompleteRequest:
      type: object
      required:
        - completed
      properties:
        completed:
          type: boolean
          description: New completion status for the task
          example: true

  responses:
    BadRequest:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: "Validation error occurred"
              error_code:
                type: string
                example: "VALIDATION_ERROR"

    Unauthorized:
      description: Unauthorized - missing or invalid authentication token
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: "Not authenticated"
              error_code:
                type: string
                example: "UNAUTHORIZED"

    Forbidden:
      description: Forbidden - user not authorized to access this resource
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: "Access denied"
              error_code:
                type: string
                example: "FORBIDDEN"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: "Task not found"
              error_code:
                type: string
                example: "NOT_FOUND"